Move parse_ollama_parameters into providers/ollama
